pyproject.toml
[project]
name = "murb-db"
version = "0.1.0"
description = "AI-ready database for Halifax RDH MURB energy model data"
requires-python = ">=3.9"
dependencies = [
    "pandas>=2.0",
    "openpyxl>=3.1",
    "pyyaml>=6.0",
    "click>=8.0",
    "matplotlib>=3.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "jupyter>=1.0",
    "ipykernel>=6.0",
    "plotly>=5.0",
]

[project.scripts]
murb-db = "murb_db.__main__:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/murb_db"]

[tool.pytest.ini_options]
testpaths = ["tests"]
